+++ package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "express": "^4.19.2",
+    "playwright": "^1.45.1"
+  }
+}
